refactor files layout for APIs v1
add a new subfolder v1 to host all api v1 related code
